+++ src/main/java/com/smartstay/smartstay/responses/assets/AssetResponse.java
@@ -1,5 +1,5 @@
 package com.smartstay.smartstay.responses.assets;
 
 
-public record AssetResponse(int id, String assetName, String brandName, String productName, String serialNumber, Double price, String purchaseDate) {
+public record AssetResponse(Long id, String assetName, String brandName, String productName, String serialNumber, Double price, String purchaseDate) {
 }
